Phase 2a-3: Extract tasks subcommand into commands/tasks/ package
Split run_tasks_command (469 lines) out of commands/__init__.py into
a focused commands/tasks/ package:

  commands/tasks/__init__.py    - dispatcher (run_tasks_command)
  commands/tasks/export.py      - was _tasks_export
  commands/tasks/push.py        - was _tasks_push
  commands/tasks/expand.py      - was _tasks_expand
  commands/tasks/visualize.py   - was _tasks_visualize
  commands/tasks/validate.py    - was _tasks_validate

run_tasks_command is re-exported from commands/__init__.py for
backwards compatibility. No behavior change — all 637 unit tests
pass.

commands/__init__.py: 1527 -> 1058 lines (-469).
