+++ src/rushti/commands/__init__.py
@@ -25,492 +25,17 @@ from rushti.app_paths import resolve_config_path
 from rushti.logging_setup import add_log_level_arg, apply_log_level
 from rushti.settings import load_settings
 from rushti.stats import create_stats_database
-from rushti.taskfile import TaskfileSource
-from rushti.taskfile_ops import (
-    analyze_runs,
-    expand_taskfile,
-    validate_taskfile_full,
-    visualize_dag,
-)
+from rushti.taskfile_ops import analyze_runs
 
 # Re-exports for backwards compatibility (Phase 2a per-subcommand splits)
 from rushti.commands.build import run_build_command  # noqa: F401
 from rushti.commands.db import run_db_command  # noqa: F401
 from rushti.commands.resume import run_resume_command  # noqa: F401
+from rushti.commands.tasks import run_tasks_command  # noqa: F401
 
 APP_NAME = "RushTI"
 
 logger = logging.getLogger()
-
-
-def run_tasks_command(argv: list) -> None:
-    """Execute the tasks command with subcommands for taskfile operations.
-
-    Subcommands:
-        export:    Convert TXT taskfiles or TM1 taskfiles to JSON format.
-        push:      Upload JSON taskfiles to TM1 as files for archival.
-        expand:    Expand MDX expressions in taskfile parameters.
-        visualize: Generate interactive HTML DAG visualization.
-        validate:  Validate taskfile structure and TM1 connectivity.
-
-    Usage:
-        # Export TXT to JSON
-        rushti tasks export --tasks tasks.txt --output tasks.json
-
-        # Export from TM1 to JSON
-        rushti tasks export --tm1-instance tm1srv01 -T DailyETL --output daily.json
-
-        # Expand MDX expressions in taskfile
-        rushti tasks expand --tasks tasks.json --output expanded.json
-
-        # Push JSON to TM1 as file
-        rushti tasks push --tasks tasks.json --tm1-instance tm1srv01
-
-        # Visualize taskfile as DAG
-        rushti tasks visualize --tasks tasks.json --output dag.html
-
-        # Validate taskfile
-        rushti tasks validate --tasks tasks.json --skip-tm1-check
-
-    :param argv: Command line arguments
-    """
-    from rushti.cli import add_taskfile_source_args
-
-    parser = argparse.ArgumentParser(
-        prog=f"{APP_NAME} tasks",
-        description="Taskfile operations: export, push, expand, visualize, and validate.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog=f"""
-Examples:
-  {APP_NAME} tasks export --tasks tasks.txt --output tasks.json
-  {APP_NAME} tasks export --tm1-instance tm1srv01 -W DailyETL --output daily.json
-  {APP_NAME} tasks push --tasks tasks.json --tm1-instance tm1srv01
-  {APP_NAME} tasks expand --tasks tasks.json --output expanded.json
-  {APP_NAME} tasks visualize --tasks tasks.json --output dag.html
-  {APP_NAME} tasks validate --tasks tasks.json --skip-tm1-check
-        """,
-    )
-
-    subparsers = parser.add_subparsers(
-        dest="subcommand",
-        required=True,
-        help="Subcommand to execute",
-    )
-
-    # --- export subcommand ---
-    export_parser = subparsers.add_parser(
-        "export",
-        help="Export taskfile to JSON format (from TXT or TM1)",
-    )
-    add_taskfile_source_args(export_parser, required=False, include_settings=True)
-    export_parser.add_argument(
-        "--output",
-        "-o",
-        dest="output_file",
-        required=True,
-        metavar="FILE",
-        help="Output JSON file path",
-    )
-    export_parser.add_argument(
-        "--mode",
-        "-m",
-        dest="mode",
-        choices=["norm", "opt"],
-        default="opt",
-        help="[TM1 sources only] 'norm' for wait-based sequencing, 'opt' for explicit predecessors. "
-        "Ignored for file sources (auto-detected). Default: opt",
-    )
-    add_log_level_arg(export_parser)
-
-    # --- push subcommand ---
-    push_parser = subparsers.add_parser(
-        "push",
-        help="Push JSON taskfile to TM1 as a file",
-    )
-    add_taskfile_source_args(push_parser, required=False, include_settings=True)
-    push_parser.add_argument(
-        "--target-tm1-instance",
-        dest="target_tm1_instance",
-        default=None,
-        metavar="INSTANCE",
-        help="Target TM1 instance for push (defaults to --tm1-instance if loading from TM1)",
-    )
-    push_parser.add_argument(
-        "--mode",
-        "-m",
-        dest="mode",
-        choices=["norm", "opt"],
-        default="opt",
-        help="[TM1 sources only] 'norm' for wait-based sequencing, 'opt' for explicit predecessors. "
-        "Ignored for file sources (auto-detected). Default: opt",
-    )
-    add_log_level_arg(push_parser)
-
-    # --- expand subcommand ---
-    expand_parser = subparsers.add_parser(
-        "expand",
-        help="Expand MDX expressions in taskfile parameters",
-    )
-    add_taskfile_source_args(expand_parser, required=False, include_settings=True)
-    expand_parser.add_argument(
-        "--output",
-        "-o",
-        dest="output_file",
-        required=True,
-        metavar="FILE",
-        help="Output file path",
-    )
-    expand_parser.add_argument(
-        "--format",
-        "-f",
-        dest="output_format",
-        choices=["json", "txt"],
-        default=None,
-        help="Output format (default: inferred from output extension)",
-    )
-    expand_parser.add_argument(
-        "--mode",
-        "-m",
-        dest="mode",
-        choices=["norm", "opt"],
-        default="opt",
-        help="[TM1 sources only] 'norm' for wait-based sequencing, 'opt' for explicit predecessors. "
-        "Ignored for file sources (auto-detected). Default: opt",
-    )
-    add_log_level_arg(expand_parser)
-
-    # --- visualize subcommand ---
-    visualize_parser = subparsers.add_parser(
-        "visualize",
-        help="Generate interactive HTML DAG visualization",
-    )
-    add_taskfile_source_args(visualize_parser, required=False, include_settings=True)
-    visualize_parser.add_argument(
-        "--output",
-        "-o",
-        dest="output_file",
-        required=True,
-        metavar="FILE",
-        help="Output HTML file path",
-    )
-    visualize_parser.add_argument(
-        "--show-parameters",
-        "-p",
-        dest="show_parameters",
-        action="store_true",
-        default=False,
-        help="Include task parameters in node labels",
-    )
-    visualize_parser.add_argument(
-        "--mode",
-        "-m",
-        dest="mode",
-        choices=["norm", "opt"],
-        default="opt",
-        help="[TM1 sources only] 'norm' for wait-based sequencing, 'opt' for explicit predecessors. "
-        "Ignored for file sources (auto-detected). Default: opt",
-    )
-    add_log_level_arg(visualize_parser)
-
-    # --- validate subcommand ---
-    validate_parser = subparsers.add_parser(
-        "validate",
-        help="Validate taskfile structure and TM1 connectivity",
-    )
-    add_taskfile_source_args(validate_parser, required=False, include_settings=True)
-    validate_parser.add_argument(
-        "--skip-tm1-check",
-        dest="skip_tm1_check",
-        action="store_true",
-        default=False,
-        help="Skip TM1 connectivity and process validation",
-    )
-    validate_parser.add_argument(
-        "--json",
-        dest="output_json",
-        action="store_true",
-        default=False,
-        help="Output results as JSON",
-    )
-    validate_parser.add_argument(
-        "--mode",
-        "-m",
-        dest="mode",
-        choices=["norm", "opt"],
-        default="opt",
-        help="[TM1 sources only] 'norm' for wait-based sequencing, 'opt' for explicit predecessors. "
-        "Ignored for file sources (auto-detected). Default: opt",
-    )
-    add_log_level_arg(validate_parser)
-
-    # Parse arguments (skip script name and 'tasks' command)
-    args = parser.parse_args(argv[2:])
-    apply_log_level(args.log_level)
-
-    # Resolve config path
-    config_path = resolve_config_path("config.ini", cli_path=None)
-
-    # Dispatch to appropriate handler based on subcommand
-    if args.subcommand == "export":
-        _tasks_export(args, config_path)
-    elif args.subcommand == "push":
-        _tasks_push(args, config_path)
-    elif args.subcommand == "expand":
-        _tasks_expand(args, config_path)
-    elif args.subcommand == "visualize":
-        _tasks_visualize(args, config_path)
-    elif args.subcommand == "validate":
-        _tasks_validate(args, config_path)
-
-
-def _tasks_export(args, config_path: str) -> None:
-    """Handle tasks --export action.
-
-    :param args: Parsed arguments
-    :param config_path: Path to config.ini
-    """
-    # Validate output is provided
-    if not args.output_file:
-        print("Error: --output is required for --export")
-        sys.exit(1)
-
-    # Build TaskfileSource from arguments
-    try:
-        source = TaskfileSource(
-            file_path=args.taskfile_path,
-            tm1_instance=args.tm1_instance,
-            workflow=args.workflow,
-        )
-        source.validate()
-    except ValueError as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-    # Validate file exists if file source
-    if source.is_file_source() and not os.path.isfile(source.file_path):
-        print(f"Error: Input file not found: {source.file_path}")
-        sys.exit(1)
-
-    try:
-        # Load taskfile from source
-        from rushti.taskfile import load_taskfile_from_source
-
-        taskfile = load_taskfile_from_source(source, config_path, mode=args.mode)
-
-        # Write to JSON
-        output_path = Path(args.output_file)
-        output_path.parent.mkdir(parents=True, exist_ok=True)
-
-        with open(output_path, "w", encoding="utf-8") as f:
-            import json
-
-            json.dump(taskfile.to_dict(), f, indent=2)
-
-        print(f"Exported {len(taskfile.tasks)} tasks to: {args.output_file}")
-        if source.is_tm1_source():
-            print(f"  Source: TM1 {source.tm1_instance}/{source.workflow}")
-        else:
-            print(f"  Source: {source.file_path}")
-        sys.exit(0)
-
-    except Exception as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-
-def _tasks_push(args, config_path: str) -> None:
-    """Handle tasks --push action.
-
-    :param args: Parsed arguments
-    :param config_path: Path to config.ini
-    """
-    # For push, we need a local file source
-    if not args.taskfile_path:
-        print("Error: --tasks is required for --push (must be a local JSON file)")
-        sys.exit(1)
-
-    if not os.path.isfile(args.taskfile_path):
-        print(f"Error: Input file not found: {args.taskfile_path}")
-        sys.exit(1)
-
-    # Determine target TM1 instance
-    target_instance = args.target_tm1_instance or args.tm1_instance
-    if not target_instance:
-        print("Error: --tm1-instance or --target-tm1-instance is required for --push")
-        sys.exit(1)
-
-    try:
-        # Load and validate the taskfile
-        from rushti.taskfile import parse_json_taskfile, detect_file_type
-
-        file_type = detect_file_type(args.taskfile_path)
-        if file_type != "json":
-            print(f"Error: --push requires a JSON taskfile, got: {file_type}")
-            sys.exit(1)
-
-        taskfile = parse_json_taskfile(args.taskfile_path)
-
-        # Connect to TM1 and upload
-        from rushti.tm1_integration import connect_to_tm1_instance
-
-        tm1 = connect_to_tm1_instance(target_instance, config_path)
-
-        try:
-            # Read file content
-            with open(args.taskfile_path, "rb") as f:
-                file_content = f.read()
-
-            # Generate filename from workflow or input filename
-            workflow = taskfile.metadata.workflow if taskfile.metadata else None
-            if not workflow:
-                workflow = Path(args.taskfile_path).stem
-
-            file_name = f"rushti_taskfile_{workflow}.json"
-
-            # Upload to TM1
-            tm1.files.create(file_name=file_name, file_content=file_content)
-
-            print(f"Pushed taskfile to TM1: {file_name}")
-            print(f"  Target instance: {target_instance}")
-            print(f"  Tasks: {len(taskfile.tasks)}")
-            sys.exit(0)
-
-        finally:
-            try:
-                tm1.logout()
-            except Exception:
-                pass  # Ignore logout errors; session may already be expired
-
-    except Exception as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-
-def _tasks_expand(args, config_path: str) -> None:
-    """Handle tasks expand action.
-
-    Expands MDX expressions in taskfile parameters and outputs a new taskfile
-    with all parameter combinations materialized.
-
-    :param args: Parsed arguments
-    :param config_path: Path to config.ini
-    """
-    # Validate output is provided
-    if not args.output_file:
-        print("Error: --output is required for expand")
-        sys.exit(1)
-
-    # Build TaskfileSource from arguments
-    try:
-        source = TaskfileSource(
-            file_path=args.taskfile_path,
-            tm1_instance=args.tm1_instance,
-            workflow=args.workflow,
-        )
-        source.validate()
-    except ValueError as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-    # Validate file exists if file source
-    if source.is_file_source() and not os.path.isfile(source.file_path):
-        print(f"Error: Input file not found: {source.file_path}")
-        sys.exit(1)
-
-    try:
-        taskfile = expand_taskfile(
-            source=source,
-            output_path=args.output_file,
-            config_path=config_path,
-            output_format=args.output_format,
-            mode=args.mode,
-        )
-        print(f"Expanded {len(taskfile.tasks)} tasks to {args.output_file}")
-        sys.exit(0)
-    except Exception as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-
-def _tasks_visualize(args, config_path: str) -> None:
-    """Handle tasks visualize action.
-
-    Generates an interactive HTML DAG visualization from a taskfile.
-
-    :param args: Parsed arguments
-    :param config_path: Path to config.ini
-    """
-    # Build TaskfileSource from arguments
-    try:
-        source = TaskfileSource(
-            file_path=args.taskfile_path,
-            tm1_instance=args.tm1_instance,
-            workflow=args.workflow,
-        )
-        source.validate()
-    except ValueError as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-    # Validate file exists if file source
-    if source.is_file_source() and not os.path.isfile(source.file_path):
-        print(f"Error: Input file not found: {source.file_path}")
-        sys.exit(1)
-
-    try:
-        result_path = visualize_dag(
-            source=source,
-            output_path=args.output_file,
-            config_path=config_path,
-            show_parameters=args.show_parameters,
-            mode=args.mode,
-        )
-        print(f"Generated DAG visualization: {result_path}")
-        sys.exit(0)
-    except Exception as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-
-def _tasks_validate(args, config_path: str) -> None:
-    """Handle tasks validate action.
-
-    Validates a taskfile structure and optionally checks TM1 connectivity.
-
-    :param args: Parsed arguments
-    :param config_path: Path to config.ini
-    """
-    # Build TaskfileSource from arguments
-    try:
-        source = TaskfileSource(
-            file_path=args.taskfile_path,
-            tm1_instance=args.tm1_instance,
-            workflow=args.workflow,
-        )
-        source.validate()
-    except ValueError as e:
-        print(f"Error: {e}")
-        sys.exit(1)
-
-    # Validate file exists if file source
-    if source.is_file_source() and not os.path.isfile(source.file_path):
-        print(f"Error: Input file not found: {source.file_path}")
-        sys.exit(1)
-
-    result = validate_taskfile_full(
-        source=source,
-        config_path=config_path,
-        check_tm1=not args.skip_tm1_check,
-        output_json=args.output_json,
-        mode=args.mode,
-    )
-
-    if args.output_json:
-        print(result.to_json())
-    else:
-        result.print_human_readable()
-
-    sys.exit(0 if result.valid else 1)
 
 
 def run_stats_command(argv: list) -> None:
